check added

+++ include/config/utils/GPUMonitorConfig.hpp
@@ -8,8 +8,8 @@ namespace height_vision_pi
     {
         double warning_temperature;
         double error_temperature;
-        int warning_usage;
-        int error_usage;
+        double warning_usage;
+        double error_usage;
 
         NLOHMANN_DEFINE_TYPE_INTRUSIVE(
             GPUMonitorConfig, warning_temperature, error_temperature, warning_usage, error_usage)
